Fixed Clang warnings.

+++ DlgEditTemplate.cpp
@@ -134,7 +134,7 @@ void DlgEditTemplate::editSelectedItem()
 		return;
 	}
 	auto row = sel[0].row();
-	auto item = m_Template.items()[row];
+	auto item = m_Template.items()[static_cast<size_t>(row)];
 	DlgEditTemplateItem dlg(m_DB, *item, this);
 	dlg.exec();
 	m_DB.saveTemplate(m_Template);
@@ -227,7 +227,7 @@ void DlgEditTemplate::cellDoubleClicked(int a_Row, int a_Column)
 {
 	Q_UNUSED(a_Column);
 
-	auto item = m_Template.items()[a_Row];
+	auto item = m_Template.items()[static_cast<size_t>(a_Row)];
 	DlgEditTemplateItem dlg(m_DB, *item, this);
 	dlg.exec();
 	m_DB.saveTemplate(m_Template);
